+++ example/main.c
@@ -3,6 +3,6 @@
 //! [ENTRYPOINT]
 int main(int argc, char* argv[])
 {
-	return cutest_run_tests(argc, argv, NULL);
+	return cutest_run_tests(argc, argv, stdout, NULL);
 }
 //! [ENTRYPOINT]