+++ example/test_f.c
@@ -1,0 +1,19 @@
+#include "cutest.h"
+
+TEST_FIXTURE_SETUP(example)
+{
+}
+
+TEST_FIXTURE_TEAREDOWN(example)
+{
+}
+
+TEST_F(example, test_f_0)
+{
+
+}
+
+TEST_F(example, test_f_1)
+{
+
+}