+++ test/src/case/list_parameterized_as_string.c
@@ -26,7 +26,7 @@ TEST_P(parameterized, as_string)
 DEFINE_TEST(parameterized, as_string, "--test_list_tests")
 {
     const char* pattern;
-    string_matrix_t* matrix = string_matrix_create_from_file(_TEST.hook.out, "");
+    string_matrix_t* matrix = string_matrix_create_from_file(_TEST.out, "");
 
     pattern = "  as_string/0  # <const char*> \"hello\"";
     ASSERT_STRING_EQ(string_matrix_access(matrix, 1, 0), pattern);