+++ test/src/case/simple.c
@@ -34,5 +34,5 @@ TEST(simple, simple)
 
 DEFINE_TEST(simple, test)
 {
-    assert(s_sum == 55);
+    CUTEST_PORTING_ASSERT(s_sum == 55);
 }