+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 2. Avoid namespace affect
 3. Able to compare any type using custom type system
 4. Use `--test_list_types` to list support types
+5. Add porting layer
 
 ### BREAKING CHANGES
 1. Hide unused function